// src/types/software/amazon/event/ruler/MatchType.ts
/**
 * The types of value matches that Ruler supports
 */
export enum MatchType {
    EXACT = 'EXACT',               // exact string
    ABSENT = 'ABSENT',              // absent key pattern
    EXISTS = 'EXISTS',              // existence pattern
    PREFIX = 'PREFIX',              // string prefix
    PREFIX_EQUALS_IGNORE_CASE = 'PREFIX_EQUALS_IGNORE_CASE', // case-insensitive string prefix
    SUFFIX = 'SUFFIX',              // string suffix
    SUFFIX_EQUALS_IGNORE_CASE = 'SUFFIX_EQUALS_IGNORE_CASE', // case-insensitive string suffix
    NUMERIC_EQ = 'NUMERIC_EQ',          // exact numeric match
    NUMERIC_RANGE = 'NUMERIC_RANGE',       // numeric range with high & low bound & </<=/>/>= options
    ANYTHING_BUT = 'ANYTHING_BUT',        // deny list effect
    ANYTHING_BUT_IGNORE_CASE = 'ANYTHING_BUT_IGNORE_CASE', // deny list effect (case insensitive)
    ANYTHING_BUT_PREFIX = 'ANYTHING_BUT_PREFIX', // anything that doesn't start with this
    ANYTHING_BUT_SUFFIX = 'ANYTHING_BUT_SUFFIX', // anything that doesn't end with this
    ANYTHING_BUT_WILDCARD = 'ANYTHING_BUT_WILDCARD', // anything that doesn't match the wildcard pattern
    EQUALS_IGNORE_CASE = 'EQUALS_IGNORE_CASE',  // case-insensitive string match
    WILDCARD = 'WILDCARD',             // string match using one or more non-consecutive '*' wildcard characters
} 